+++ src/plugins/recordTypes/collectionobject/forms/timebased.jsx
@@ -56,6 +56,13 @@ const template = (configContext) => {
             </Field>
 
             <Field name="computedCurrentLocation" />
+
+            <Field name="homeLocationGroupList">
+              <Field name="homeLocationGroup">
+                <Field name="homeLocation" />
+                <Field name="homeLocationNote" />
+              </Field>
+            </Field>
           </Col>
         </Row>